% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawproj-package.R
\docType{package}
\name{rawproj-package}
\alias{rawproj}
\alias{rawproj-package}
\title{rawproj: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/rawproj/}
  \item \url{https://github.com/rmgpanw/rawproj}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID})

}
\keyword{internal}
