% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renviron_to_list}
\alias{renviron_to_list}
\title{Read .Renviron file to list}
\usage{
renviron_to_list(path, suffix = "_RENV")
}
\arguments{
\item{path}{Path to directory containing a \code{.Renviron} file.}

\item{suffix}{Optional suffix to append to names of objects extracted from
\code{.Renviron}.}
}
\value{
A named list.
}
\description{
Reads environmental variables defined by a \code{.Renviron} file to a named list.
}
\examples{
# create `.Renviron` in `tempdir()` and read to current R session
temp_dotrenviron <- file.path(tempdir(), ".Renviron")
cat("KEY=value", file = temp_dotrenviron)
readRenviron(temp_dotrenviron)

# get environmental variables specified by `.Renviron` to a named list
renviron_to_list(path = tempdir())
}
